vtkParticleTracerBase:: Improve functionality for coprocessing
1. Add a flag to disable this->ResetCache()
2. Remove an unnecessary condition when advancing this->CurrentTimeStep.
   This condition causes extra time requests in a coprocessing work flow
3. More graceful handling of the situation when there is only one time step
4. Fixes a few warnings and stylistic issues.

+++ Filters/FlowPaths/Testing/Cxx/TestParticleTracers.cxx
@@ -56,7 +56,7 @@ public:
   }
   int GetNumberOfTimeSteps()
   {
-    return this->TimeSteps.size();
+    return static_cast<int>(this->TimeSteps.size());
   }
 
 protected:
@@ -131,7 +131,7 @@ protected:
                  range,2);
 
     outInfo->Set(vtkStreamingDemandDrivenPipeline::TIME_STEPS(),
-                 &TimeSteps[0], TimeSteps.size());
+                 &TimeSteps[0], static_cast<int>(TimeSteps.size()));
 
 
     outInfo->Set(vtkStreamingDemandDrivenPipeline::WHOLE_EXTENT(), this->Extent,6);
